refactor: Assign and unassign flow
