--- src/app/interactions/componentInteractions/handlers/changeRole/index.ts
@@ -1,2 +1,0 @@
-export * from './assignRoleHandler';
-export * from './unassignRoleHandler';
