+++ src/app/interactions/componentInteractions/handlers/configuration/index.ts
@@ -18,3 +18,9 @@ export * from './3.3_handleConfirmAlertsToSend';
 export * from './3.4_handleFrequencyOfAlertsToSend';
 export * from './3_handleSendAlerts';
 export * from './4_handleFinalMessage';
+export * from './handleAddUserCircle';
+export * from './handleAddUserCircleConfirmation';
+export * from './handleAddUserCircleCancel';
+export * from './handleRemoveUserCircle';
+export * from './handleRemoveUserCircleConfirmation';
+export * from './handleRemoveUserCircleCancel';
