+++ src/app/interactions/customId.ts
@@ -4,6 +4,8 @@ export enum CustomId {
     AllowAlertsButton = 'ALLOW_ALERTS_BUTTON',
     AssignButton = 'ASSIGN_BUTTON',
     AssignRoleUserSelect = 'ASSIGN_ROLE_USER_SELECT',
+    AssignRoleUserSelectConfirmButton = 'ASSIGN_ROLE_USER_SELECT_CONFIRM_BUTTON',
+    AssignRoleUserSelectCancelButton = 'ASSIGN_ROLE_USER_SELECT_CANCEL_BUTTON',
     CircleSelect = 'CIRCLE_SELECT',
     CircleSelectNextButton = 'CIRCLE_SELECT_NEXT_BUTTON',
     ConfigButton = 'CONFIGURE_BUTTON',
@@ -14,9 +16,10 @@ export enum CustomId {
     LinkedChannelAlertsUpdateButton = 'LINKED_CHANNEL_ALERTS_UPDATE_BUTTON',
     Skip = 'SKIP',
     SkipLinkCircleButton = 'SKIP_LINK_CIRCLE_BUTTON',
+    UnlinkButton = 'UNLINK_BUTTON',
     UnassignButton = 'UNASSIGN_BUTTON',
     UnassignRoleUserSelect = 'UNASSIGN_ROLE_USER_SELECT',
-    UnlinkButton = 'UNLINK_BUTTON',
-    UnssignRoleUserSelect = 'UNASSIGN_ROLE_USER_SELECT',
+    UnassignRoleUserSelectConfirmButton = 'UNASSIGN_ROLE_USER_SELECT_CONFIRM_BUTTON',
+    UnassignRoleUserSelectCancelButton = 'UNASSIGN_ROLE_USER_SELECT_CANCEL_BUTTON',
     UpdateAlertsButton = 'UPDATE_ALERTS_BUTTON',
 }
